rename testing_util.go to testing_util_stubs_oss.go and move init BuildDate to new shared testing_util.go
